initial implementation of runtime config building
This code is... not pleasant.

Runtime Config translation from kubernetes-native api through to
spin-native api is rather complex when accounting for the number of
places secrets and config maps can be referenced.

This commit attempts to at least shove all of the nastier bits into a
single package that can be _mostly_ ignored.

Signed-off-by: Danielle Lancashire <dani@builds.terrible.systems>