Executor: Make runtimeClass configurable
Move the definition of the underlying `runtimeClass` to the executor
config. This allows an operator to specify a runtimeClass name that
makes sense for their cluster, and allows for multiple runtimeClasses
for trying out different node pool versions in a live cluster.

Also move deployment creation to the CRD to remove cyclotron-specific
handling.

Signed-off-by: Danielle Lancashire <dani@builds.terrible.systems>