adds indexing for spec.executor field
To query/list a custom resource by a field in spec,
it must be indexed by the controller executing the query.
This adds indexing for spec.executor so that upon delete,
we can filter and list spinapps by spec.executor.
This resolves the behavior we were seeing around not being
able to delete a spinappexecutor even after spinapp workloads
were deleted.

resolves #102

Signed-off-by: Michelle Dhanani <michelle@fermyon.com>