runtimeconfig: Fix a Panic when an empty value is desired
Previously if value was `""` we would panic because we would always
attempt to load ValueFrom, which can be nil. This is partially a design
bug: Empty values aren't _really_ expected to be set. - But Spin _will_
allow them, so we should technically let them be specified?

This allows them to be specified by being a fallback value when nothing
else is available, and avoids the panic, which is a bug regardless.

Signed-off-by: Danielle Lancashire <dani@builds.terrible.systems>