ca-secrets: Annotate secrets that are created by spin-operator
Eventually we will want to upgrade caBundles to new versions as
certificates expire/are rotated/new ca's form. We should however only do
that when we manage the resource that is created.

Here we introduce a managed-by annotation that we can use to ensure that
we don't replace user-managed secret bundles in the future.